@@ -27,6 +27,7 @@ use Keboola\JobQueueInternalClient\JobFactory;
 use Keboola\JobQueueInternalClient\JobFactory\JobInterface;
 use Keboola\JobQueueInternalClient\JobFactory\JobResult;
 use Keboola\JobQueueInternalClient\JobPatchData;
+use Keboola\ObjectEncryptor\Exception\UserException as EncryptionUserException;
 use Keboola\StorageApi\Client as StorageClient;
 use Keboola\StorageApiBranch\ClientWrapper;
 use Monolog\Logger;
@@ -38,29 +39,17 @@ use Throwable;
 
 class RunCommand extends Command
 {
-    /** @var string */
+    /**
+     * @inheritdoc
+     */
     protected static $defaultName = 'app:run';
-
-    /** @var string */
-    private $legacyOauthApiUrl;
-
-    /** @var array */
-    private $instanceLimits;
-
-    /** @var QueueClient */
-    private $queueClient;
-
-    /** @var Logger */
-    private $logger;
-
-    /** @var LogProcessor */
-    private $logProcessor;
-
-    /** @var StorageApiFactory */
-    private $storageApiFactory;
-
-    /** @var JobDefinitionFactory */
-    private $jobDefinitionFactory;
+    private string $legacyOauthApiUrl;
+    private array $instanceLimits;
+    private QueueClient $queueClient;
+    private Logger $logger;
+    private LogProcessor $logProcessor;
+    private StorageApiFactory $storageApiFactory;
+    private JobDefinitionFactory $jobDefinitionFactory;
 
     public function __construct(
         LoggerInterface $logger,
@@ -73,7 +62,11 @@ class RunCommand extends Command
     ) {
         parent::__construct(self::$defaultName);
         $this->queueClient = $queueClient;
-        $this->logger = $logger;
+        if (is_a($logger, Logger::class)) {
+            $this->logger = $logger;
+        } else {
+            $this->logger = new Logger('not-used');
+        }
         $this->storageApiFactory = $storageApiFactory;
         $this->jobDefinitionFactory = $jobDefinitionFactory;
         $this->legacyOauthApiUrl = $legacyOauthApiUrl;
@@ -183,7 +176,7 @@ class RunCommand extends Command
             }
             $this->logger->info(sprintf('Job "%s" execution finished.', $jobId));
             $this->postJobResult($jobId, JobFactory::STATUS_SUCCESS, $result);
-        } catch (\Keboola\ObjectEncryptor\Exception\UserException $e) {
+        } catch (EncryptionUserException $e) {
             $this->logger->error(
                 sprintf('Job "%s" ended with encryption error: "%s"', $jobId, $e->getMessage()),
                 ExceptionTransformer::transformException($e)->getFullArray()
