@@ -11,23 +11,13 @@ use Keboola\JobQueueInternalClient\Exception\ClientException;
 use Keboola\Temp\Temp;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Filesystem\Filesystem;
+use function GuzzleHttp\json_encode;
 
 class UsageFileTest extends TestCase
 {
-    /**
-     * @var Temp
-     */
-    private $temp;
-
-    /**
-     * @var string
-     */
-    private $dataDir;
-
-    /**
-     * @var Filesystem
-     */
-    private $fs;
+    private Temp $temp;
+    private string $dataDir;
+    private Filesystem $fs;
 
     public function setUp(): void
     {
@@ -48,7 +38,7 @@ class UsageFileTest extends TestCase
     public function testStoreUsageWrongDataJson(): void
     {
         // there should be "metric" key instead of "random"
-        $usage = \GuzzleHttp\json_encode([[
+        $usage = json_encode([[
             'random' => 'API calls',
             'value' => 150,
         ]]);
@@ -95,7 +85,7 @@ YAML;
 
     public function testStoreUsageOk(): void
     {
-        $usage = \GuzzleHttp\json_encode([[
+        $usage = json_encode([[
             'metric' => 'kiloBytes',
             'value' => 150,
         ]]);
@@ -125,7 +115,7 @@ YAML;
 
     public function testStoreUsageUnknownJob(): void
     {
-        $usage = \GuzzleHttp\json_encode([[
+        $usage = json_encode([[
             'metric' => 'kiloBytes',
             'value' => 150,
         ]]);
@@ -153,7 +143,7 @@ YAML;
 
     public function testDoNotStoreEmptyUsage(): void
     {
-        $usage = \GuzzleHttp\json_encode([]);
+        $usage = json_encode([]);
         $this->fs->dumpFile($this->dataDir . '/out/usage.json', $usage);
 
         $client = self::getMockBuilder(Client::class)
